refactor once more
