+++ ui/snippets/navigation/vertical/__screenshots__/NavigationDesktop.pw.tsx_default_hover-dark-mode-1.png
